+++ android/anc/src/main/java/org/smartregister/fhircore/anc/ui/anccare/encounters/EncounterDataProvider.kt
@@ -18,7 +18,7 @@ package org.smartregister.fhircore.anc.ui.anccare.encounters
 
 import androidx.paging.PagingData
 import kotlinx.coroutines.flow.Flow
-import org.smartregister.fhircore.anc.data.anc.model.EncounterItem
+import org.smartregister.fhircore.anc.data.model.EncounterItem
 
 interface EncounterDataProvider {