+++ android/anc/src/main/java/org/smartregister/fhircore/anc/ui/details/vitalsigns/VitalSignsDetailsViewModel.kt
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2021 Ona Systems, Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.smartregister.fhircore.anc.ui.details.vitalsigns
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import kotlinx.coroutines.launch
+import org.smartregister.fhircore.anc.data.model.EncounterItem
+import org.smartregister.fhircore.anc.data.patient.PatientRepository
+import org.smartregister.fhircore.engine.util.DefaultDispatcherProvider
+import org.smartregister.fhircore.engine.util.DispatcherProvider
+
+class VitalSignsDetailsViewModel(
+  val ancPatientRepository: PatientRepository,
+  var dispatcher: DispatcherProvider = DefaultDispatcherProvider,
+  val patientId: String
+) : ViewModel() {
+
+  fun fetchEncounters(): LiveData<List<EncounterItem>> {
+    val patientEncounters = MutableLiveData<List<EncounterItem>>()
+    viewModelScope.launch(dispatcher.io()) {
+      val listEncounters = ancPatientRepository.fetchEncounters(patientId = patientId)
+      val listEncountersItem = arrayListOf<EncounterItem>()
+      patientEncounters.postValue(listEncountersItem)
+    }
+    return patientEncounters
+  }
+}