+++ android/anc/src/test/java/org/smartregister/fhircore/anc/ui/details/PatientDetailsActivityTest.kt
@@ -1,0 +1,107 @@
+/*
+ * Copyright 2021 Ona Systems, Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.smartregister.fhircore.anc.ui.details
+
+import android.app.Activity
+import android.view.MenuInflater
+import androidx.arch.core.executor.testing.InstantTaskExecutorRule
+import com.google.android.fhir.FhirEngine
+import io.mockk.coEvery
+import io.mockk.every
+import io.mockk.mockk
+import io.mockk.spyk
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import org.junit.Assert
+import org.junit.Before
+import org.junit.Rule
+import org.junit.Test
+import org.junit.jupiter.api.DisplayName
+import org.robolectric.Robolectric
+import org.robolectric.annotation.Config
+import org.smartregister.fhircore.anc.activity.ActivityRobolectricTest
+import org.smartregister.fhircore.anc.coroutine.CoroutineTestRule
+import org.smartregister.fhircore.anc.data.model.AncPatientDetailItem
+import org.smartregister.fhircore.anc.data.model.AncPatientItem
+import org.smartregister.fhircore.anc.data.patient.PatientRepository
+import org.smartregister.fhircore.anc.shadow.AncApplicationShadow
+import org.smartregister.fhircore.anc.ui.anccare.details.AncDetailsViewModel
+
+@ExperimentalCoroutinesApi
+@Config(shadows = [AncApplicationShadow::class])
+internal class PatientDetailsActivityTest : ActivityRobolectricTest() {
+
+  private lateinit var patientDetailsActivity: PatientDetailsActivity
+
+  private lateinit var patientDetailsActivitySpy: PatientDetailsActivity
+
+  private lateinit var fhirEngine: FhirEngine
+
+  private lateinit var patientDetailsViewModel: AncDetailsViewModel
+
+  private lateinit var patientRepository: PatientRepository
+
+  @get:Rule var coroutinesTestRule = CoroutineTestRule()
+
+  @get:Rule var instantTaskExecutorRule = InstantTaskExecutorRule()
+
+  private val patientId = "samplePatientId"
+  var ancPatientDetailItem = spyk<AncPatientDetailItem>()
+
+  @Before
+  fun setUp() {
+
+    fhirEngine = mockk(relaxed = true)
+
+    patientRepository = mockk()
+
+    every { ancPatientDetailItem.patientDetails } returns
+      AncPatientItem(patientId, "Mandela Nelson", "M", "26")
+    every { ancPatientDetailItem.patientDetailsHead } returns AncPatientItem()
+    coEvery { patientRepository.fetchDemographics(patientId) } returns ancPatientDetailItem
+
+    patientDetailsViewModel =
+      spyk(
+        AncDetailsViewModel(patientRepository, coroutinesTestRule.testDispatcherProvider, patientId)
+      )
+
+    patientDetailsActivity =
+      Robolectric.buildActivity(PatientDetailsActivity::class.java, null).create().get()
+    patientDetailsActivitySpy = spyk(objToCopy = patientDetailsActivity)
+  }
+
+  @Test
+  @DisplayName("Should start patient details activity")
+  fun testPatientActivityShouldNotNull() {
+    Assert.assertNotNull(patientDetailsActivity)
+  }
+
+  @Test
+  @DisplayName("Should inflate menu and return true")
+  fun testThatMenuIsCreated() {
+
+    val menuInflater = mockk<MenuInflater>()
+
+    every { patientDetailsActivitySpy.menuInflater } returns menuInflater
+    every { menuInflater.inflate(any(), any()) } returns Unit
+
+    Assert.assertTrue(patientDetailsActivitySpy.onCreateOptionsMenu(null))
+  }
+
+  override fun getActivity(): Activity {
+    return patientDetailsActivity
+  }
+}